# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: OpenaiMessage[];
}

export interface ConversationInput {
  title: string;
}

export interface MessageInput {
  content: string;
}

export interface VoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export type ImageGenerationInputStyle =
  (typeof ImageGenerationInputStyle)[keyof typeof ImageGenerationInputStyle];

export const ImageGenerationInputStyle = {
  realistic: "realistic",
  anime: "anime",
  logo: "logo",
  artistic: "artistic",
} as const;

export type ImageGenerationInputSize =
  (typeof ImageGenerationInputSize)[keyof typeof ImageGenerationInputSize];

export const ImageGenerationInputSize = {
  "1024x1024": "1024x1024",
  "1536x1024": "1536x1024",
  "1024x1536": "1024x1536",
} as const;

export interface ImageGenerationInput {
  prompt: string;
  style?: ImageGenerationInputStyle;
  size?: ImageGenerationInputSize;
}

export interface ImageGenerationResult {
  b64_json: string;
  prompt: string;
}

export interface HumanizeInput {
  text: string;
}

export interface HumanizeResult {
  humanized: string;
  original: string;
}

export interface YoutubeInput {
  url: string;
  summarize?: boolean;
}

export interface YoutubeResult {
  transcript: string;
  /** @nullable */
  summary: string | null;
  videoId: string;
  /** @nullable */
  title: string | null;
}

export type CodeInputMode = (typeof CodeInputMode)[keyof typeof CodeInputMode];

export const CodeInputMode = {
  generate: "generate",
  debug: "debug",
  explain: "explain",
} as const;

export interface CodeInput {
  prompt: string;
  language: string;
  mode: CodeInputMode;
  /** @nullable */
  code?: string | null;
}

export interface CodeResult {
  result: string;
  language: string;
  mode: string;
}

export interface SongInput {
  mood: string;
  genre: string;
  language: string;
  /** @nullable */
  theme?: string | null;
}

export interface SongResult {
  lyrics: string;
  title: string;
  mood: string;
  genre: string;
}

export interface PdfExportInput {
  conversationId: number;
}

export interface PdfExportResult {
  filename: string;
  /** Base64-encoded PDF content */
  content: string;
}

export interface OpenaiError {
  error: string;
}