# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: tools
    description: AI tool endpoints (humanizer, youtube, code, song, image-to-video)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceMessageInput"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}

  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerationInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationResult"

  /tools/humanize:
    post:
      operationId: humanizeText
      tags: [tools]
      summary: Convert AI-generated text to human-like text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanizeInput"
      responses:
        "200":
          description: Humanized text result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanizeResult"

  /tools/youtube-transcript:
    post:
      operationId: getYoutubeTranscript
      tags: [tools]
      summary: Get transcript from a YouTube video URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YoutubeInput"
      responses:
        "200":
          description: YouTube transcript and summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeResult"

  /tools/generate-code:
    post:
      operationId: generateCode
      tags: [tools]
      summary: Generate, debug, or explain code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: Generated/explained code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeResult"

  /tools/generate-song:
    post:
      operationId: generateSong
      tags: [tools]
      summary: Generate song lyrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "200":
          description: Generated song lyrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongResult"

  /tools/export-pdf:
    post:
      operationId: exportConversationPdf
      tags: [tools]
      summary: Export conversation as PDF (returns base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfExportInput"
      responses:
        "200":
          description: PDF export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfExportResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
        - updatedAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - updatedAt
        - messages

    ConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    MessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    VoiceMessageInput:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio

    ImageGenerationInput:
      type: object
      properties:
        prompt:
          type: string
        style:
          type: string
          enum: [realistic, anime, logo, artistic]
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
      required:
        - prompt

    ImageGenerationResult:
      type: object
      properties:
        b64_json:
          type: string
        prompt:
          type: string
      required:
        - b64_json
        - prompt

    HumanizeInput:
      type: object
      properties:
        text:
          type: string
      required:
        - text

    HumanizeResult:
      type: object
      properties:
        humanized:
          type: string
        original:
          type: string
      required:
        - humanized
        - original

    YoutubeInput:
      type: object
      properties:
        url:
          type: string
        summarize:
          type: boolean
      required:
        - url

    YoutubeResult:
      type: object
      properties:
        transcript:
          type: string
        summary:
          type: ["string", "null"]
        videoId:
          type: string
        title:
          type: ["string", "null"]
      required:
        - transcript
        - summary
        - videoId
        - title

    CodeInput:
      type: object
      properties:
        prompt:
          type: string
        language:
          type: string
        mode:
          type: string
          enum: [generate, debug, explain]
        code:
          type: ["string", "null"]
      required:
        - prompt
        - language
        - mode

    CodeResult:
      type: object
      properties:
        result:
          type: string
        language:
          type: string
        mode:
          type: string
      required:
        - result
        - language
        - mode

    SongInput:
      type: object
      properties:
        mood:
          type: string
        genre:
          type: string
        language:
          type: string
        theme:
          type: ["string", "null"]
      required:
        - mood
        - genre
        - language

    SongResult:
      type: object
      properties:
        lyrics:
          type: string
        title:
          type: string
        mood:
          type: string
        genre:
          type: string
      required:
        - lyrics
        - title
        - mood
        - genre

    PdfExportInput:
      type: object
      properties:
        conversationId:
          type: integer
      required:
        - conversationId

    PdfExportResult:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
          description: Base64-encoded PDF content
      required:
        - filename
        - content

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error