# lib/api-zod/src/generated/types/codeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CodeInputMode } from "./codeInputMode";

export interface CodeInput {
  prompt: string;
  language: string;
  mode: CodeInputMode;
  /** @nullable */
  code?: string | null;
}