# lib/api-zod/src/generated/types/imageGenerationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageGenerationInputSize } from "./imageGenerationInputSize";
import type { ImageGenerationInputStyle } from "./imageGenerationInputStyle";

export interface ImageGenerationInput {
  prompt: string;
  style?: ImageGenerationInputStyle;
  size?: ImageGenerationInputSize;
}