# lib/api-zod/src/generated/types/pdfExportResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PdfExportResult {
  filename: string;
  /** Base64-encoded PDF content */
  content: string;
}