# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  mood: string;
  genre: string;
  language: string;
  /** @nullable */
  theme?: string | null;
}