# lib/api-zod/src/generated/types/voiceMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}