# lib/api-zod/src/generated/types/youtubeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface YoutubeResult {
  transcript: string;
  /** @nullable */
  summary: string | null;
  videoId: string;
  /** @nullable */
  title: string | null;
}